Initial commit of the De Boor algorithm for spline interpolation
A `Spline' class was created and equipped with a recursive method for
generating the basis functions. These functions can be created by using the
overwritten `__call__' method of the class or by using the `knot_sequence'
method directly. The former was implemented in the main block to plot two
such basis functions.