Implementation of de Boor- and bloosom algorithms for computng B-splines
The SplineClass script contains a `Spline' class which can be used to
translate points in the plane to de Boor points (and vice versa) using a
vandermonde matrix of basis functions. The spline is the calculated using
the __call__ method, where a number of points (resolution) is specified. The
result can be visualized using a combination of of four plot options.

The script `SplineDemo' was written to show tha capabilities of the class,
and the `SplineTests' script contains unittests for tha class to test if
everything has been downloaded correctly.